split TavueTable from Tavue component
